+++ src-bindings/Exception/BindingsNotCollected.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Ray\Di\Exception;
+namespace Ray\Bindings\Exception;
 
 use LogicException;
 
 /** Bindings were rendered before a module snapshot was collected. */
-final class BindingsNotCollected extends LogicException implements ExceptionInterface
+final class BindingsNotCollected extends LogicException
 {
 }
